pyproject.toml
[project]
name = "bias_check"
version = "0.2.0"
description = "Wrapping a bias check model into a validator."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0",
    "transformers>=4.40.2",
    "tensorflow>=2.18.0"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
] 

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "tests"
]

[tool.pyright]
include = ["validator"]
